+++ src/HHbbVV/postprocessing/templates/23Oct3VBF/lpsfs.csv
@@ -1,0 +1,12 @@
+,SF,syst_unc,stat_unc,sj_pt_unc,sj_matching_unc
+HHbbVV,0.93 ± 0.21,0.08735995513326773,0.02178232885498517,0.0061149325206791465,0.21056243132864816
+ggHH_kl_2p45_kt_1_HHbbVV,0.94 ± 0.22,0.09825588332373836,0.017728326791750713,0.007576634540588289,0.2109042985659727
+ggHH_kl_5_kt_1_HHbbVV,0.90 ± 0.31,0.26670664011549117,0.029146162489653592,0.009392097264437689,0.21495771560903995
+ggHH_kl_0_kt_1_HHbbVV,0.97 ± 0.22,0.0855042203073505,0.019319390242510937,0.0057113506172409585,0.21249833138903884
+VBFHHbbVV,0.97 ± 0.29,0.21606146805271081,0.022866755372481395,0.011791744840525328,0.19970070579826676
+qqHH_CV_1_C2V_0_kl_1_HHbbVV,0.95 ± 0.18,0.0440400535017882,0.016305841465033288,0.05912205378108328,0.1762753439506925
+qqHH_CV_1p5_C2V_1_kl_1_HHbbVV,0.97 ± 0.19,0.043170552991571264,0.01843468279444452,0.05548925971951003,0.17659871392784582
+qqHH_CV_1_C2V_1_kl_2_HHbbVV,0.93 ± 0.41,0.3741327061402456,0.0365511621336087,0.015097514340344167,0.2173996175908222
+qqHH_CV_1_C2V_2_kl_1_HHbbVV,0.96 ± 0.18,0.04751934156410037,0.016217262585498455,0.06330571028260944,0.17501290974191908
+qqHH_CV_1_C2V_1_kl_0_HHbbVV,0.97 ± 0.22,0.12492900211693345,0.017097719425056068,0.00947242206235012,0.1944825092307107
+qqHH_CV_0p5_C2V_1_kl_1_HHbbVV,0.95 ± 0.18,0.043193276036633484,0.01618427553643857,0.06082940744062617,0.1750516645939186
