+++ src/HHbbVV/postprocessing/templates/23Oct3VBF/systematics.json
@@ -1,0 +1,122 @@
+{
+    "2017": {
+        "pass": {
+            "trig_total": 87.6234207605359,
+            "trig_total_err": 1.232538971320039
+        },
+        "fail": {
+            "trig_total": 346.5607033736205,
+            "trig_total_err": 0.7244257707820206
+        }
+    },
+    "HHbbVV": {
+        "lp_sf": 0.9255566155790668,
+        "lp_sf_unc": 0.2290854894968195,
+        "lp_sf_uncs": {
+            "syst_unc": 0.08735995513326773,
+            "stat_unc": 0.02178232885498517,
+            "sj_pt_unc": 0.0061149325206791465,
+            "sj_matching_unc": 0.21056243132864816
+        }
+    },
+    "ggHH_kl_2p45_kt_1_HHbbVV": {
+        "lp_sf": 0.9391472823461275,
+        "lp_sf_unc": 0.23346635886810285,
+        "lp_sf_uncs": {
+            "syst_unc": 0.09825588332373836,
+            "stat_unc": 0.017728326791750713,
+            "sj_pt_unc": 0.007576634540588289,
+            "sj_matching_unc": 0.2109042985659727
+        }
+    },
+    "ggHH_kl_5_kt_1_HHbbVV": {
+        "lp_sf": 0.8985676068767847,
+        "lp_sf_unc": 0.34391417775434757,
+        "lp_sf_uncs": {
+            "syst_unc": 0.26670664011549117,
+            "stat_unc": 0.029146162489653592,
+            "sj_pt_unc": 0.009392097264437689,
+            "sj_matching_unc": 0.21495771560903995
+        }
+    },
+    "ggHH_kl_0_kt_1_HHbbVV": {
+        "lp_sf": 0.9670162916725116,
+        "lp_sf_unc": 0.22993992889167636,
+        "lp_sf_uncs": {
+            "syst_unc": 0.0855042203073505,
+            "stat_unc": 0.019319390242510937,
+            "sj_pt_unc": 0.0057113506172409585,
+            "sj_matching_unc": 0.21249833138903884
+        }
+    },
+    "VBFHHbbVV": {
+        "lp_sf": 0.9679521376909226,
+        "lp_sf_unc": 0.2953385576267809,
+        "lp_sf_uncs": {
+            "syst_unc": 0.21606146805271081,
+            "stat_unc": 0.022866755372481395,
+            "sj_pt_unc": 0.011791744840525328,
+            "sj_matching_unc": 0.19970070579826676
+        }
+    },
+    "qqHH_CV_1_C2V_0_kl_1_HHbbVV": {
+        "lp_sf": 0.9481126483385646,
+        "lp_sf_unc": 0.19176501481383776,
+        "lp_sf_uncs": {
+            "syst_unc": 0.0440400535017882,
+            "stat_unc": 0.016305841465033288,
+            "sj_pt_unc": 0.05912205378108328,
+            "sj_matching_unc": 0.1762753439506925
+        }
+    },
+    "qqHH_CV_1p5_C2V_1_kl_1_HHbbVV": {
+        "lp_sf": 0.971884691277887,
+        "lp_sf_unc": 0.19097041100787895,
+        "lp_sf_uncs": {
+            "syst_unc": 0.043170552991571264,
+            "stat_unc": 0.01843468279444452,
+            "sj_pt_unc": 0.05548925971951003,
+            "sj_matching_unc": 0.17659871392784582
+        }
+    },
+    "qqHH_CV_1_C2V_1_kl_2_HHbbVV": {
+        "lp_sf": 0.93214579437633,
+        "lp_sf_unc": 0.43451328854827137,
+        "lp_sf_uncs": {
+            "syst_unc": 0.3741327061402456,
+            "stat_unc": 0.0365511621336087,
+            "sj_pt_unc": 0.015097514340344167,
+            "sj_matching_unc": 0.2173996175908222
+        }
+    },
+    "qqHH_CV_1_C2V_2_kl_1_HHbbVV": {
+        "lp_sf": 0.9569683205759278,
+        "lp_sf_unc": 0.19276467248739185,
+        "lp_sf_uncs": {
+            "syst_unc": 0.04751934156410037,
+            "stat_unc": 0.016217262585498455,
+            "sj_pt_unc": 0.06330571028260944,
+            "sj_matching_unc": 0.17501290974191908
+        }
+    },
+    "qqHH_CV_1_C2V_1_kl_0_HHbbVV": {
+        "lp_sf": 0.9679770635328802,
+        "lp_sf_unc": 0.23197577622646612,
+        "lp_sf_uncs": {
+            "syst_unc": 0.12492900211693345,
+            "stat_unc": 0.017097719425056068,
+            "sj_pt_unc": 0.00947242206235012,
+            "sj_matching_unc": 0.1944825092307107
+        }
+    },
+    "qqHH_CV_0p5_C2V_1_kl_1_HHbbVV": {
+        "lp_sf": 0.9529558040722325,
+        "lp_sf_unc": 0.1909735373189062,
+        "lp_sf_uncs": {
+            "syst_unc": 0.043193276036633484,
+            "stat_unc": 0.01618427553643857,
+            "sj_pt_unc": 0.06082940744062617,
+            "sj_matching_unc": 0.1750516645939186
+        }
+    }
+}
